# docs/swagger.yaml
openapi: 3.0.3
info:
  title: BondRota API
  description: Documentação completa da API BondRota para consumo do frontend, extraída
    de `docs/api-reference.md`.
  version: 1.0.0
servers:
- url: https://bondrota-api.onrender.com/api/v1
  description: Servidor de Produção
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: bondrota_admin_session
    planningCronAuth:
      type: http
      scheme: bearer
      description: Segredo interno configurado em PLANNING_CRON_SECRET
  schemas:
...
    ClientePublicId:
      type: string
      pattern: '^cli_[A-Za-z0-9_-]{21}$'
      example: cli_N8f2Qx5mL0sV7kWp3ZaBc
...
    VinculoPublicId:
      type: string
      pattern: '^vin_[A-Za-z0-9_-]{21}$'
      example: vin_H3mQ8sTx1cV6kNp0ZaB7d
...
    CreateClienteRequest:
      type: object
      required:
      - nome
      - cpf
      - senha
      - data_nasc
      - documento_identificacao
      - comprovante_residencia
      properties:
        nome:
          type: string
          example: Maria Cliente
        cpf:
          type: string
          description: Aceita 11 digitos ou a mascara 000.000.000-00.
          example: '11111111111'
        senha:
          type: string
          example: senha123
        telefone:
          type: string
          description: Aceita 11 digitos ou a mascara (00) 00000-0000.
          example: '82999991111'
        data_nasc:
          type: string
          format: date
          example: '2002-08-10'
        documento_identificacao:
          type: string
          description: Path privado de PDF, JPG, JPEG, PNG ou WebP no bucket documentos.
          example: clientes/_novo/uuid/documento-identificacao.pdf
        comprovante_residencia:
          type: string
          description: Path privado de PDF, JPG, JPEG, PNG ou WebP no bucket documentos.
          example: clientes/_novo/uuid/comprovante-residencia.pdf
    UpdateClienteRequest:
      type: object
      properties:
        nome:
          type: string
          example: Maria Cliente
        telefone:
          type: string
          nullable: true
          description: Aceita 11 digitos ou a mascara (00) 00000-0000. Omitido/null preserva o valor atual; string vazia limpa o campo.
          example: '82999991111'
        data_nasc:
          type: string
          format: date
          example: '2002-08-10'
        documento_identificacao:
          type: string
          nullable: true
          description: Omitido/null preserva o valor atual; string vazia e rejeitada.
          example: clientes/1/documento-identificacao.pdf
        comprovante_residencia:
          type: string
          nullable: true
          description: Omitido/null preserva o valor atual; string vazia e rejeitada.
          example: clientes/1/comprovante-residencia.pdf
    ClienteResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ClientePublicId'
        nome:
          type: string
          example: Maria Cliente
        cpf:
          type: string
          example: '11111111111'
        telefone:
          type: string
          example: '82999991111'
        data_nasc:
          type: string
          format: date
          example: '2002-08-10'
        documento_identificacao:
          type: string
          example: clientes/1/documento-identificacao.pdf
        comprovante_residencia:
          type: string
          example: clientes/1/comprovante-residencia.pdf
    ClienteListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClienteResponse'
        next_cursor:
          type: string
          description: Opaco. So presente quando has_more e true.
          example: MTIw
        has_more:
          type: boolean
...
    VinculoHorarioFixo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        vinculo_id:
          $ref: '#/components/schemas/VinculoPublicId'
        dia_semana:
          type: integer
          example: 1
    VinculoResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/VinculoPublicId'
        cliente_id:
          $ref: '#/components/schemas/ClientePublicId'
        tipo:
          type: string
          enum:
          - estudante
          - estagio
          example: estudante
        turno:
          type: string
          enum:
          - MT
          - VT
          - NT
          - IN
          example: NT
        destino_id:
          type: integer
          example: 1
        rota_interna_id:
          type: integer
          example: 1
        curso:
          type: string
          example: Sistemas de Informacao
        comprovante:
          type: string
          example: https://...
        validade:
          type: string
          format: date
          example: '2026-12-31'
        horarios_fixos:
          type: array
          items:
            $ref: '#/components/schemas/VinculoHorarioFixo'
...
    ClienteComVinculosResponse:
      allOf:
      - $ref: '#/components/schemas/ClienteResponse'
      - type: object
        properties:
          vinculos:
            type: array
            items:
              $ref: '#/components/schemas/VinculoResponse'
...
security:
- bearerAuth: []
- adminCookieAuth: []
paths:
...
  /clientes/:
    post:
      summary: Cria cliente
      tags:
      - Clientes
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '409':
          description: CPF ou telefone já pertence a outro cadastro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClienteRequest'
    get:
      summary: Lista clientes, paginada por cursor
      description: >-
        q busca por nome, telefone ou CPF. A busca por CPF ignora pontuacao, mas
        so quando o termo nao tem letra — senao "Ana 13" tambem procuraria
        documentos contendo "13".
      tags:
      - Clientes
      parameters:
      - name: cursor
        in: query
        required: false
        description: Opaco, devolvido em next_cursor.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          maximum: 200
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteListResponse'
...
  /clientes/{clienteID}:
    get:
      summary: Busca cliente com vinculos
      tags:
      - Clientes
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteComVinculosResponse'
      parameters:
      - name: clienteID
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ClientePublicId'
    put:
      summary: Atualiza cliente
      tags:
      - Clientes
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '409':
          description: Telefone já pertence a outro cadastro
      parameters:
      - name: clienteID
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ClientePublicId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClienteRequest'
    delete:
      summary: Remove cliente
      tags:
      - Clientes
      responses:
        '204':
          description: No Content
        '409':
          description: Registro em uso por outro cadastro
      parameters:
      - name: clienteID
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ClientePublicId'
...